pier14/opc-client: Tweaks

+++ pier14/opc-client/press.sh
@@ -7,10 +7,18 @@ echo "Ready"
 while true
 do
 	read -s i
-	echo -n "Pressed "
+	if test "$i" = "r"
+	then
+		echo "Restarting"
+		sudo systemctl restart opc-client.service
+		continue
+	fi
+
+	echo -n "$(date +%H:%M:%S)  Pressed "
 	touch /var/run/soma/buttonA
 
-	read -s i
+	#read -s i
+	sleep 0.5
 	echo "Released"
 	rm -f /var/run/soma/buttonA
 done
